make proxied composed input params composite/methods place refactor
